+++ middleware.ts
@@ -1,6 +1,5 @@
-import { match } from '@formatjs/intl-localematcher';
-import Negotiator from 'negotiator';
 import { NextRequest, NextResponse } from 'next/server';
+import { saveLocaleToCookieServer, getLocale } from 'utils/locale';
 
 export enum SupportedLocales {
   EN = 'en',
@@ -33,52 +32,12 @@ export function middleware(request: NextRequest) {
     url.pathname = `/${locale}${pathname}`;
     response = NextResponse.redirect(url);
   }
-  response.cookies.set('NEXT_LOCALE', locale, {
-    maxAge: 60 * 60 * 24 * 365,
-    path: '/',
-  });
-  return response;
+  return saveLocaleToCookieServer(response, locale);
 }
 
 export const config = {
   matcher: ['/'],
 };
-
-function getLocale(request: NextRequest): string {
-  const cookieLocale = getLocaleFromCookie(request);
-  if (cookieLocale) {
-    return cookieLocale;
-  }
-
-  const acceptLanguageLocale = getLocaleFromAcceptLanguage(request);
-  if (acceptLanguageLocale) {
-    return acceptLanguageLocale;
-  }
-
-  return DEFAULT_LOCALE;
-}
-
-function getLocaleFromCookie(request: NextRequest): string | null {
-  const cookieLocale = request.cookies.get('NEXT_LOCALE')?.value;
-
-  if (cookieLocale && SUPPORTED_LOCALES.includes(cookieLocale)) {
-    return cookieLocale;
-  }
-
-  return null;
-}
-
-function getLocaleFromAcceptLanguage(request: NextRequest): string | null {
-  const negotiatorHeaders: Record<string, string> = {};
-  request.headers.forEach((value, key) => (negotiatorHeaders[key] = value));
-
-  let languages = new Negotiator({ headers: negotiatorHeaders }).languages();
-  try {
-    return match(languages, SUPPORTED_LOCALES, DEFAULT_LOCALE);
-  } catch (error) {
-    return null;
-  }
-}
 
 export function isValidPathCheck(pathname) {
   if (VALID_PATHS.includes(pathname)) {
